+++ src/pypi_status.py
@@ -1,0 +1,3 @@
+"""Backward-compatible exports for PyPI publication helpers."""
+
+from .interfaces.pypi_status import *  # noqa: F401,F403